requirements.txt
Flask>=1.0.0
Flask-SQLAlchemy==2.3.2
PyMySQL==0.9.2
requests[security]==2.20.1
sqlalchemy==1.2.18
gunicorn>=19.5.0
google-cloud-storage
google-cloud-vision
httplib2
git+git://github.com/GoogleCloudPlatform/psq@master
oauth2client
honcho